move additonal new query code to internal
